+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Fixed
 - Fixed ref forwarding support for the `ListItem` component, allowing access to the underlying DOM element.
+- Fixed the issue where the Autocomplete tooltip was overwritten by transient input.
 
 ### Changed